Update ConvertTZ trait to improve performance
Replaced the constructor with a model's retrieved event for mutating datetime attributes in ConvertTZ trait. This change improves performance as mutation is now performed only when necessary, i.e., when the model is retrieved from the database. The construction caused unnecessary mutations by being invoked multiple times. Method 'generateAccessorName' was also removed as it was no longer required with this new approach.